+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 
+## Version 0.2.0 (2023-10-25)
+### Bugs
+- Fixed logging handlers
+- Removed hardcoded timeout for waiting for threads
+
 
 ## Version 0.0.4 (2023-04-08)